docker compose: Add volume mounts
This allows to change the config and to view logs outside of the container by
creating mount points on the local filesystem. The directories must exist and
be owned by the default user (uid=1000,gid=1000) so that they can be mounted
by the container on startup. Paths and UIDs to be amended as needed.